task-cli/site/events-dav/requirements.txt
flask
requests
opencv-python-headless
numpy
insightface
Pillow
onnxruntime 